requirements.txt
APScheduler==3.10.1
Flask==2.3.0
Flask_Cors==3.0.10
psycopg2_binary==2.9.6
Requests==2.29.0
Flask[async]
aiohttp